Maintenance Mode & Update readme gifs
I have been mulling this over in my head for a few weeks. I have decided to only focus on maintenance requests for the indefinite future because I have to budget my time. Youtube Tabs is in a good place. There are certainly improvements that could be made and further implementation ideas that I already have (and had previously planned). Maybe those will come sometime in the far future when/if I have more time available to me.